+++ apps/jubmoji-quest/src/hooks/useFetchPowers.ts
@@ -6,12 +6,57 @@ import {
 import { JubmojiPower } from "@/types";
 import { Jubmoji, ProvingState } from "jubmoji-api";
 import { useMutation, useQuery } from "react-query";
-
+import { useJubmojis } from "./useJubmojis";
+import { $Enums, Prisma } from "@prisma/client";
 interface PowerMutationProps {
   power: JubmojiPower;
   jubmojis: Jubmoji[];
   onUpdateProvingState?: (provingState: ProvingState) => void;
 }
+
+export const useGetPowerLockedStatus = (
+  powerId: string | number | undefined
+) => {
+  const { data: jubmojis = [] } = useJubmojis();
+  const { data: power } = useFetchPowerById(powerId);
+
+  return useQuery(
+    ["questPowerLockedStatus", power?.id, jubmojis.length],
+    async (): Promise<{ locked: boolean }> => {
+      if (!power) {
+        return {
+          locked: true,
+        };
+      }
+
+      const collectionCardIndices = power.collectionCards.map(
+        (card) => card.index
+      );
+
+      const collectedItems =
+        jubmojis?.filter((jubmoji) =>
+          collectionCardIndices.includes(jubmoji.pubKeyIndex)
+        ).length ?? 0;
+
+      const proofParams = power.proofParams as Prisma.JsonObject;
+
+      if (power.proofType === $Enums.ProofType.N_UNIQUE_IN_COLLECTION) {
+        return {
+          locked: collectedItems < (proofParams.N as number),
+        };
+      }
+
+      return {
+        locked: collectionCardIndices.length > 0 && collectedItems === 0,
+      };
+    },
+    {
+      refetchOnWindowFocus: false,
+      enabled: power?.id !== undefined,
+      retry: power?.id !== undefined,
+    }
+  );
+};
 
 export const useFetchPowers = () => {
   return useQuery(
